requirements.txt
mesop
google-genai
python-pptx
python-docx
EbookLib
beautifulsoup4
python-dotenv
